src/osfv/libs/models.py: list_models command, YML loading moved to Models
Models class introduced, model data (*.yml) loading moved from RTE to
Models. load_model_data() method modified to act as *.yml validator.
list_models top-level command introduced to CLI.
