+++ osfv_cli/test/data/FakeDeviceBroken.yml
@@ -1,0 +1,14 @@
+---
+flash_chip:
+  model: "W25Q64JV-.Q"
+  voltage: "3.3V"
+
+frogrammer:
+  name: rte_1_0
+
+pwr_ctrl:
+  sonoff: false
+  relay: true
+  flashing_power_state: "S5"
+
+reset_cmos: true
